Put getSystemInfoSync in systemInfo.
