+++ pal/screen-adapter/minigame/screen-adapter.ts
@@ -57,40 +57,6 @@ try {
     console.error(e);
 }
 
-const originalGetSystemInfoSync = minigame.getSystemInfoSync;
-let _cachedSystemInfo: SystemInfo = originalGetSystemInfoSync.call(minigame);
-
-function testAndUpdateSystemInfoCache (testAmount: number, testInterval: number): void {
-    let successfullyTestTimes = 0;
-    let intervalTimer: number | null = null;
-    function testCachedSystemInfo (): void {
-        const currentSystemInfo = originalGetSystemInfoSync.call(minigame);
-        if (_cachedSystemInfo.screenWidth === currentSystemInfo.screenWidth && _cachedSystemInfo.screenHeight === currentSystemInfo.screenHeight) {
-            if (++successfullyTestTimes >= testAmount && intervalTimer !== null) {
-                clearInterval(intervalTimer);
-                intervalTimer = null;
-            }
-        } else {
-            successfullyTestTimes = 0;
-        }
-        _cachedSystemInfo = currentSystemInfo;
-    }
-    intervalTimer = setInterval(testCachedSystemInfo, testInterval);
-}
-
-if (WECHAT) {
-    testAndUpdateSystemInfoCache(10, 500);
-}
-
-minigame.onWindowResize?.(() => {
-    // update cached system info
-    _cachedSystemInfo = originalGetSystemInfoSync.call(minigame);
-});
-
-minigame.getSystemInfoSync = function (): SystemInfo {
-    return _cachedSystemInfo;
-};
-
 class ScreenAdapter extends EventTarget {
     public isFrameRotated = false;
     public handleResizeEvent = true;
