+++ pal/system-info/minigame/system-info.ts
@@ -23,7 +23,7 @@
 */
 
 import { ALIPAY, BYTEDANCE, HUAWEI, OPPO, RUNTIME_BASED, VIVO, MIGU, HONOR, WECHAT, XIAOMI, DEBUG, TEST, TAOBAO, TAOBAO_MINIGAME, WECHAT_MINI_PROGRAM } from 'internal:constants';
-import { minigame } from 'pal/minigame';
+import { minigame, SystemInfo as MinigameSystenInfo } from 'pal/minigame';
 import { IFeatureMap } from 'pal/system-info';
 import { EventTarget } from '../../../cocos/core/event';
 import { checkPalIntegrity, withImpl } from '../../integrity-check';
@@ -92,6 +92,41 @@ if (BYTEDANCE) {
         return true;
     };
 }
+
+const originalGetSystemInfoSync = minigame.getSystemInfoSync;
+let _cachedSystemInfo: MinigameSystenInfo = originalGetSystemInfoSync.call(minigame);
+
+function testAndUpdateSystemInfoCache (testAmount: number, testInterval: number): void {
+    let successfullyTestTimes = 0;
+    let intervalTimer: number | null = null;
+    function testCachedSystemInfo (): void {
+        const currentSystemInfo = originalGetSystemInfoSync.call(minigame);
+        if (_cachedSystemInfo.screenWidth === currentSystemInfo.screenWidth && _cachedSystemInfo.screenHeight === currentSystemInfo.screenHeight) {
+            if (++successfullyTestTimes >= testAmount && intervalTimer !== null) {
+                clearInterval(intervalTimer);
+                intervalTimer = null;
+            }
+        } else {
+            successfullyTestTimes = 0;
+        }
+        _cachedSystemInfo = currentSystemInfo;
+    }
+    intervalTimer = setInterval(testCachedSystemInfo, testInterval);
+}
+
+if (WECHAT) {
+    testAndUpdateSystemInfoCache(10, 500);
+}
+
+minigame.onWindowResize?.(() => {
+    // update cached system info
+    _cachedSystemInfo = originalGetSystemInfoSync.call(minigame);
+});
+
+minigame.getSystemInfoSync = function (): MinigameSystenInfo {
+    return _cachedSystemInfo;
+};
+
 class SystemInfo extends EventTarget {
     public declare readonly networkType: NetworkType;
     public declare readonly isNative: boolean;
